Add ScenarioDisplayConfig; move EnergyConsumptionByVehicleTypeAnalyzer to generic module
- Add ScenarioDisplayConfig dataclass to framework for generic scenario
  ordering, display names, and baseline selection
- Move EnergyConsumptionByVehicleTypeAnalyzer from bvg_tools.py to
  output_analyzers.py as a data-only analyzer (no visualize method)
- Move merge_energy_consumption_results() alongside the analyzer
- Add BVGEnergyConsumptionAnalyzer subclass in bvg_tools.py with
  LaTeX-styled table visualization
- Add optional config parameter to merge_scenario_comparisons(),
  merge_energy_consumption_results(), and visualize_tco_comparison()
  for generic scenario ordering (falls back to hardcoded BVG defaults)
