+++ eflips/x/steps/analyzers/output_analyzers.py
@@ -18,7 +18,16 @@ import sqlalchemy
 from eflips.eval.output import prepare as eval_output_prepare
 from eflips.eval.output import visualize as eval_output_visualize
 from eflips.eval.output.prepare import depot_layout
-from eflips.model import EnergySource, Scenario, VehicleType
+from eflips.model import (
+    EnergySource,
+    Event,
+    EventType,
+    Route,
+    Scenario,
+    Trip,
+    Vehicle,
+    VehicleType,
+)
 from eflips.tco.data_queries import init_tco_parameters  # type: ignore[import-untyped]
 from eflips.tco.tco_calculator import TCOCalculator  # type: ignore[import-untyped]
 from eflips.tco.tco_parameter_config import (  # type: ignore[import-untyped]
@@ -30,7 +39,7 @@ from eflips.tco.tco_parameter_config import (  # type: ignore[import-untyped]
 )
 from sqlalchemy.orm import Session
 
-from eflips.x.framework import Analyzer
+from eflips.x.framework import Analyzer, ScenarioDisplayConfig
 
 logger = logging.getLogger(__name__)
 
@@ -977,3 +986,114 @@ def merge_tco_results(results: List[pd.DataFrame]) -> pd.DataFrame:
         Combined DataFrame with all scenarios
     """
     return pd.concat(results, ignore_index=True)
+
+
+# ============================================================================
+# Energy Consumption and Battery-Electric Range
+# ============================================================================
+
+
+class EnergyConsumptionByVehicleTypeAnalyzer(Analyzer):
+    """
+    Compute average energy consumption and battery-electric range per vehicle type and scenario.
+
+    For each driving event in the simulation results, calculates energy consumed from
+    ``(soc_start - soc_end) * battery_capacity``, then aggregates per vehicle type to give
+    average consumption in kWh/km and a derived battery-electric range.
+    """
+
+    def __init__(self, code_version: str = "v1.0.0", cache_enabled: bool = True):
+        super().__init__(code_version=code_version, cache_enabled=cache_enabled)
+
+    @classmethod
+    def document_params(cls) -> Dict[str, str]:
+        return {
+            f"{cls.__name__}.scenario_name": "Label for this scenario in the output table.",
+        }
+
+    def analyze(self, session: Session, params: Dict[str, Any]) -> pd.DataFrame:
+        scenario_name = params.get(f"{self.__class__.__name__}.scenario_name", "Unknown")
+
+        rows = (
+            session.query(Event, VehicleType, Route)
+            .join(Vehicle, Event.vehicle_id == Vehicle.id)
+            .join(VehicleType, Vehicle.vehicle_type_id == VehicleType.id)
+            .join(Trip, Event.trip_id == Trip.id)
+            .join(Route, Trip.route_id == Route.id)
+            .filter(Event.event_type == EventType.DRIVING)
+            .all()
+        )
+
+        records = []
+        for event, vtype, route in rows:
+            energy_kwh = (event.soc_start - event.soc_end) * vtype.battery_capacity
+            distance_km = route.distance / 1000
+            records.append(
+                {
+                    "vehicle_type_id": vtype.id,
+                    "vehicle_type": vtype.name,
+                    "vehicle_type_short": vtype.name_short,
+                    "battery_capacity_kwh": vtype.battery_capacity,
+                    "battery_capacity_reserve_kwh": vtype.battery_capacity_reserve,
+                    "energy_kwh": energy_kwh,
+                    "distance_km": distance_km,
+                }
+            )
+
+        df = pd.DataFrame(records)
+
+        result_rows = []
+        for vtype_id, group in df.groupby("vehicle_type_id"):
+            total_energy = group["energy_kwh"].sum()
+            total_distance = group["distance_km"].sum()
+            avg_consumption = total_energy / total_distance if total_distance > 0 else float("nan")
+            battery_cap = group["battery_capacity_kwh"].iloc[0]
+            battery_reserve = group["battery_capacity_reserve_kwh"].iloc[0]
+            usable_battery = battery_cap - battery_reserve
+            range_km = usable_battery / avg_consumption if avg_consumption > 0 else float("nan")
+
+            result_rows.append(
+                {
+                    "scenario_name": scenario_name,
+                    "vehicle_type": group["vehicle_type"].iloc[0],
+                    "vehicle_type_short": group["vehicle_type_short"].iloc[0],
+                    "avg_consumption_kwh_per_km": round(avg_consumption, 2),
+                    "battery_capacity_kwh": battery_cap,
+                    "usable_battery_kwh": usable_battery,
+                    "battery_electric_range_km": round(range_km, 0),
+                }
+            )
+
+        result_rows.sort(key=lambda r: r["vehicle_type_short"])
+        return pd.DataFrame(result_rows)
+
+
+def merge_energy_consumption_results(
+    dfs: List[pd.DataFrame],
+    config: "ScenarioDisplayConfig | None" = None,
+) -> pd.DataFrame:
+    """
+    Merge single-scenario energy consumption DataFrames into one table.
+
+    Args:
+        dfs: List of DataFrames from ``EnergyConsumptionByVehicleTypeAnalyzer.analyze()``.
+        config: Optional scenario display configuration for ordering.
+                Falls back to hardcoded ``["OU", "DEP", "TERM"]`` order if not provided.
+
+    Returns:
+        Combined DataFrame sorted by scenario order then vehicle type.
+    """
+    merged = pd.concat(dfs, ignore_index=True)
+
+    if config is not None:
+        merged["_scenario_sort"] = merged["scenario_name"].map(lambda s: config.sort_key(s))
+    else:
+        scenario_order = {name: i for i, name in enumerate(["OU", "DEP", "TERM"])}
+        merged["_scenario_sort"] = merged["scenario_name"].map(scenario_order).fillna(99)
+    merged = (
+        merged.sort_values(["_scenario_sort", "vehicle_type_short"])
+        .drop(columns="_scenario_sort")
+        .reset_index(drop=True)
+    )
+
+    return merged
